+++ src/main/java/sensors/TemperatureSensor.java
@@ -4,37 +4,37 @@ import controllers.TemperatureController;
 
 public class TemperatureSensor {
 
-    private static final double HIGH_TEMPERATURE_THRESHOLD = 104.0;
-    private static final double LOW_TEMPERATURE_THRESHOLD = 50.0;
-    private static final double OPTIMAL_TEMPERATURE = 77.0;
+    private static final int HIGH_TEMPERATURE_THRESHOLD = 104;
+    private static final int LOW_TEMPERATURE_THRESHOLD = 50;
+    private static final int OPTIMAL_TEMPERATURE = 77;
 
-    private static double temperature = OPTIMAL_TEMPERATURE; // Default temperature
+    private static int temperature = OPTIMAL_TEMPERATURE; // Default temperature
 
-    public static double getTemperature() {
+    public static int getTemperature() {
         return temperature;
     }
 
-    public static void updateTemperature(double newTemperature) {
+    public static void updateTemperature(int newTemperature) {
         temperature = newTemperature;
         temperature = checkAndAdjustTemperature();
     }
 
-    private static double checkAndAdjustTemperature() {
+    private static int checkAndAdjustTemperature() {
         if (temperature > HIGH_TEMPERATURE_THRESHOLD || temperature < LOW_TEMPERATURE_THRESHOLD) {
             temperature = TemperatureController.adjustTemperature(temperature);
         }
         return temperature;
     }
 
-    public static double getHighTemperatureThreshold() {
+    public static int getHighTemperatureThreshold() {
         return HIGH_TEMPERATURE_THRESHOLD;
     }
 
-    public static double getLowTemperatureThreshold() {
+    public static int getLowTemperatureThreshold() {
         return LOW_TEMPERATURE_THRESHOLD;
     }
 
-    public static double getOptimalTemperature() {
+    public static int getOptimalTemperature() {
         return OPTIMAL_TEMPERATURE;
     }
 }
